Cleaner project Deleted unnecessary files and renamed others
