+++ bar-trender-front-end/src/components/Modals/ModalLogin.js
@@ -3,7 +3,7 @@ import React from "react";
 
 import { Button, Modal, ModalBody } from "reactstrap";
 
-import POSTLoginForm from "./ApiLoginForm";
+import POSTLoginForm from "../ApiLoginForm";
 
 function ModalLogin() {
   const [modal1, setModal1] = React.useState(false);
