--- bar-trender-front-end/src/components/Navbars/ExamplesNavbar.js
@@ -1,138 +1,0 @@
-import React from "react";
-import barTrender from '../../assets/img/barTrender60.png';
-// reactstrap components
-import {
-  Collapse,
-  NavbarBrand,
-  Navbar,
-  NavItem,
-  NavLink,
-  Nav,
-  Container,
-  UncontrolledTooltip,
-  FormGroup,
-  Input,
-} from "reactstrap";
-
-import ModalSearch from "../../components/ModalSearch";
-
-
-function ExamplesNavbar() {
-  const [navbarColor, setNavbarColor] = React.useState("navbar-transparent");
-  const [collapseOpen, setCollapseOpen] = React.useState(false);
-  React.useEffect(() => {
-    const updateNavbarColor = () => {
-      if (
-        document.documentElement.scrollTop > 399 ||
-        document.body.scrollTop > 399
-      ) {
-        setNavbarColor("");
-      } else if (
-        document.documentElement.scrollTop < 400 ||
-        document.body.scrollTop < 400
-      ) {
-        setNavbarColor("navbar-transparent");
-      }
-    };
-    window.addEventListener("scroll", updateNavbarColor);
-    return function cleanup() {
-      window.removeEventListener("scroll", updateNavbarColor);
-    };
-  });
-  return (
-    <>
-      {collapseOpen ? (
-        <div
-          id="bodyClick"
-          onClick={() => {
-            document.documentElement.classList.toggle("nav-open");
-            setCollapseOpen(false);
-          }}
-        />
-      ) : null}
-      <Navbar className={"fixed-top " + navbarColor}  color="info" expand="lg">
-        <Container>
-
-          <div className="navbar-translate">
-            <Nav>
-              <NavLink className="Logo">
-                <img  alt="" src={barTrender}  />
-              </NavLink>
-              <NavbarBrand
-                href="/landing-page"
-                target="_blank"
-                id="navbar-brand"
-              >
-                BarTrender
-                </NavbarBrand>
-              <UncontrolledTooltip target="#navbar-brand">
-                BarTrender - LandingPage
-                </UncontrolledTooltip>
-              <button
-                className="navbar-toggler navbar-toggler"
-                onClick={() => {
-                  document.documentElement.classList.toggle("nav-open");
-                  setCollapseOpen(!collapseOpen);
-                }}
-                aria-expanded={collapseOpen}
-                type="button"
-              >
-                <span className="navbar-toggler-bar top-bar"></span>
-                <span className="navbar-toggler-bar middle-bar"></span>
-                <span className="navbar-toggler-bar bottom-bar"></span>
-              </button>
-            </Nav>
-          </div>
-          <div>
-            <FormGroup>
-              <Input
-                defaultValue=""
-                placeholder="Regular"
-                type="text"
-              ></Input>
-            </FormGroup> </div>
-
-          <ModalSearch />
-
-          <Collapse
-            className="justify-content-end"
-            isOpen={collapseOpen}
-            navbar
-          >
-            <Nav navbar>
-              <NavItem>
-                <NavLink
-                  href="https://twitter.com/TrenderBar"
-                  target="_blank"
-                  id="twitter-tooltip"
-                >
-                  <i className="fab fa-twitter"></i>
-                  <p className="d-lg-none d-xl-none">Twitter</p>
-                </NavLink>
-                <UncontrolledTooltip target="#twitter-tooltip">
-                  Síguenos en Twitter
-                </UncontrolledTooltip>
-              </NavItem>
-
-              <NavItem>
-                <NavLink
-                  href="https://www.instagram.com/bartrenderofficial/"
-                  target="_blank"
-                  id="instagram-tooltip"
-                >
-                  <i className="fab fa-instagram"></i>
-                  <p className="d-lg-none d-xl-none">Instagram</p>
-                </NavLink>
-                <UncontrolledTooltip target="#instagram-tooltip">
-                  Síguenos en Instagram
-                </UncontrolledTooltip>
-              </NavItem>
-            </Nav>
-          </Collapse>
-        </Container>
-      </Navbar>
-    </>
-  );
-}
-
-export default ExamplesNavbar;
