+++ bar-trender-front-end/src/index.js
@@ -14,14 +14,13 @@ import "./assets/scss/now-ui-kit.scss?v=1.4.0";
 import "./assets/demo/demo.css?v=1.4.0";
 import "./assets/demo/nucleo-icons-page-styles.css?v=1.4.0";
 // pages for this kit
-
-import MainPage from "./views/MainPage.js";
-import LandingPage from "./views/LandingPage.js";
-import ListPage from "./views/FilterResults.js"
-import ErrorPage from "./views/ErrorPage.js"
+import MainView from "./views/MainView.js";
+import LandingView from "./views/LandingView.js";
+import ListView from "./views/ListView.js"
+import ErrorView from "./views/ErrorView.js"
 
 const app = (
-
+  
   <React.StrictMode>
     <BrowserRouter>
     <Switch>
@@ -29,27 +28,24 @@ const app = (
         <Route path="/app" render={(props) => <App {...props} />} />
        
         <Route
-          path="/index"
-          render={(props) => <LandingPage {...props} />}
+          path="/landing"
+          render={(props) => <LandingView {...props} />}
         />
         <Route
           path="/main"
-
-          render={(props) => <MainPage {...props} />}
-
+          render={(props) => <MainView {...props} />}
         />
         <Route
           path="/list"
-          render={(props) => <ListPage {...props} />}
+          render={(props) => <ListView {...props} />}
         />
-
         <Route
           path="/error"
-          render={(props) => <ErrorPage {...props} />}
+          render={(props) => <ErrorView {...props} />}
         />
-
-        <Redirect to="/index" />
-        <Redirect from="/" to="/index" />
+       
+        <Redirect to="/main" />
+        <Redirect from="/" to="/main" />
       </Switch>
     </Switch>
   </BrowserRouter>
