+++ bar-trender-front-end/src/views/ErrorView.js
@@ -33,7 +33,7 @@ import employee_9 from '../assets/img/miguel.png';
 import employee_10 from '../assets/img/miguel-angel.png';
 import employee_11 from '../assets/img/fran.png';
 
-function LandingPage() {
+function ErrorView() {
   const [firstFocus, setFirstFocus] = React.useState(false);
   const [lastFocus, setLastFocus] = React.useState(false);
   React.useEffect(() => {
@@ -70,4 +70,4 @@ function LandingPage() {
   );
 }
 
-export default LandingPage;
+export default ErrorView;
