+++ bar-trender-front-end/src/views/LandingView.js
@@ -33,7 +33,7 @@ import employee_9 from '../assets/img/miguel.png';
 import employee_10 from '../assets/img/miguel-angel.png';
 import employee_11 from '../assets/img/fran.png';
 
-function LandingPage() {
+function LandingView() {
   const [firstFocus, setFirstFocus] = React.useState(false);
   const [lastFocus, setLastFocus] = React.useState(false);
   React.useEffect(() => {
@@ -369,4 +369,4 @@ function LandingPage() {
   );
 }
 
-export default LandingPage;
+export default LandingView;
