--- bar-trender-front-end/src/views/index-sections/BasicElements.js
@@ -1,295 +1,0 @@
-import React from "react";
-// react plugin used to create switch buttons
-import Switch from "react-bootstrap-switch";
-// plugin that creates slider
-import Slider from "nouislider";
-
-// reactstrap components
-import {
-  Button,
-  Label,
-  FormGroup,
-  Input,
-  InputGroupAddon,
-  InputGroupText,
-  InputGroup,
-  Container,
-  Row,
-  Col,
-} from "reactstrap";
-
-// core components
-
-function BasicElements() {
-  const [leftFocus, setLeftFocus] = React.useState(false);
-  const [rightFocus, setRightFocus] = React.useState(false);
-  React.useEffect(() => {
-    if (
-      !document
-        .getElementById("sliderRegular")
-        .classList.contains("noUi-target")
-    ) {
-      Slider.create(document.getElementById("sliderRegular"), {
-        start: [50],
-        connect: [true, false],
-        step: 0.5,
-        range: { min: 0, max: 100 },
-      });
-    }
-    if (
-      !document.getElementById("sliderDouble").classList.contains("noUi-target")
-    ) {
-      Slider.create(document.getElementById("sliderDouble"), {
-        start: [20, 80],
-        connect: [false, true, false],
-        step: 1,
-        range: { min: 0, max: 100 },
-      });
-    }
-  });
-  return (
-    <>
-      <div className="section section-basic" id="basic-elements">
-        <Container>
-          <h3 className="title">Basic Elements</h3>
-          <h4>Buttons</h4>
-          <p className="category">Pick your style</p>
-          <Row>
-            <Col md="10">
-              <Button color="info" type="button">
-                Default
-              </Button>
-              <Button className="btn-round" color="info" type="button">
-                Round
-              </Button>
-              <Button className="btn-round" color="info" type="button">
-                <i className="now-ui-icons ui-2_favourite-28"></i>
-                With Icon
-              </Button>
-              <Button className="btn-icon btn-round" color="info" type="button">
-                <i className="now-ui-icons ui-2_favourite-28"></i>
-              </Button>
-              <Button className="btn-round" color="info" outline type="button">
-                Outline
-              </Button>
-            </Col>
-          </Row>
-          <p className="category">Pick your size</p>
-          <Row>
-            <Col md="10">
-              <Button color="info" size="sm">
-                Small
-              </Button>
-              <Button color="info">Regular</Button>
-              <Button color="info" size="lg">
-                Large
-              </Button>
-            </Col>
-          </Row>
-          <p className="category">Pick your color</p>
-          <Row>
-            <Col md="10">
-              <Button color="default">Default</Button>
-              <Button color="primary">Primary</Button>
-              <Button color="info">Info</Button>
-              <Button color="success">Success</Button>
-              <Button color="warning">Warning</Button>
-              <Button color="danger">Danger</Button>
-              <Button className="btn-neutral" color="default">
-                Neutral
-              </Button>
-            </Col>
-          </Row>
-          <h4>Links</h4>
-          <Row>
-            <Col md="8">
-              <Button color="link">Default</Button>
-              <Button className="btn-link" color="primary">
-                Primary
-              </Button>
-              <Button className="btn-link" color="info">
-                Info
-              </Button>
-              <Button className="btn-link" color="success">
-                Success
-              </Button>
-              <Button className="btn-link" color="warning">
-                Warning
-              </Button>
-              <Button className="btn-link" color="danger">
-                Danger
-              </Button>
-            </Col>
-          </Row>
-          <div className="space-70"></div>
-          <div id="inputs">
-            <h4>Inputs</h4>
-            <p className="category">Form Controls</p>
-            <Row>
-              <Col lg="3" sm="6">
-                <FormGroup>
-                  <Input
-                    defaultValue=""
-                    placeholder="Regular"
-                    type="text"
-                  ></Input>
-                </FormGroup>
-              </Col>
-              <Col lg="3" sm="6">
-                <FormGroup className="has-success">
-                  <Input
-                    className="form-control-success"
-                    defaultValue="Success"
-                    type="text"
-                  ></Input>
-                </FormGroup>
-              </Col>
-              <Col lg="3" sm="6">
-                <FormGroup className="has-danger">
-                  <Input
-                    className="form-control-danger"
-                    defaultValue="Error Input"
-                    type="email"
-                  ></Input>
-                </FormGroup>
-              </Col>
-              <Col lg="3" sm="6">
-                <InputGroup className={leftFocus ? "input-group-focus" : ""}>
-                  <InputGroupAddon addonType="prepend">
-                    <InputGroupText>
-                      <i className="fa fa-user-circle"></i>
-                    </InputGroupText>
-                  </InputGroupAddon>
-                  <Input
-                    placeholder="Left Font Awesome Icon"
-                    type="text"
-                    onFocus={() => setLeftFocus(true)}
-                    onBlur={() => setLeftFocus(false)}
-                  ></Input>
-                </InputGroup>
-              </Col>
-              <Col lg="3" sm="6">
-                <InputGroup className={rightFocus ? "input-group-focus" : ""}>
-                  <Input
-                    placeholder="Right Nucleo Icon"
-                    type="text"
-                    onFocus={() => setRightFocus(true)}
-                    onBlur={() => setRightFocus(false)}
-                  ></Input>
-                  <InputGroupAddon addonType="append">
-                    <InputGroupText>
-                      <i className="now-ui-icons users_single-02"></i>
-                    </InputGroupText>
-                  </InputGroupAddon>
-                </InputGroup>
-              </Col>
-            </Row>
-          </div>
-          <div className="space-70"></div>
-          <Row id="checkRadios">
-            <Col className="mb-4" lg="3" sm="6">
-              <p className="category">Checkboxes</p>
-              <FormGroup check>
-                <Label check>
-                  <Input type="checkbox" ></Input>
-                  <span className="form-check-sign"></span>
-                  Unchecked
-                </Label>
-              </FormGroup>
-              <FormGroup check>
-                <Label check>
-                  <Input defaultChecked type="checkbox"></Input>
-                  <span className="form-check-sign"></span>
-                  Checked
-                </Label>
-              </FormGroup>
-              <FormGroup check disabled>
-                <Label check>
-                  <Input disabled type="checkbox"></Input>
-                  <span className="form-check-sign"></span>
-                  Disabled Unchecked
-                </Label>
-              </FormGroup>
-              <FormGroup check disabled>
-                <Label check>
-                  <Input defaultChecked disabled type="checkbox"></Input>
-                  <span className="form-check-sign"></span>
-                  Disabled Checked
-                </Label>
-              </FormGroup>
-            </Col>
-            <Col className="mb-4" lg="3" sm="6">
-              <p className="category">Radios</p>
-              <FormGroup check className="form-check-radio">
-                <Label check>
-                  <Input
-                    defaultValue="option1"
-                    id="exampleRadios1"
-                    name="exampleRadios"
-                    type="radio"
-                  ></Input>
-                  <span className="form-check-sign"></span>
-                  Radio is off
-                </Label>
-              </FormGroup>
-              <FormGroup check className="form-check-radio">
-                <Label check>
-                  <Input
-                    defaultChecked
-                    defaultValue="option2"
-                    id="exampleRadios1"
-                    name="exampleRadios"
-                    type="radio"
-                  ></Input>
-                  <span className="form-check-sign"></span>
-                  Radio is on
-                </Label>
-              </FormGroup>
-              <FormGroup check className="form-check-radio" disabled>
-                <Label check>
-                  <Input
-                    defaultValue="option3"
-                    disabled
-                    id="exampleRadios2"
-                    name="exampleRadios1"
-                    type="radio"
-                  ></Input>
-                  <span className="form-check-sign"></span>
-                  Disabled radio is off
-                </Label>
-              </FormGroup>
-              <FormGroup check className="form-check-radio" disabled>
-                <Label check>
-                  <Input
-                    defaultChecked
-                    defaultValue="option4"
-                    disabled
-                    id="exampleRadios2"
-                    name="exampleRadios1"
-                    type="radio"
-                  ></Input>
-                  <span className="form-check-sign"></span>
-                  Disabled radio is on
-                </Label>
-              </FormGroup>
-            </Col>
-            <Col lg="3" sm="6">
-              <p className="category">Toggle Buttons</p>
-              <Switch offColor="" offText="" onColor="" onText=""></Switch>
-              <br></br>
-              <Switch defaultValue={false} offColor="" onColor=""></Switch>
-            </Col>
-            <Col lg="3" sm="6">
-              <p className="category">Sliders</p>
-              <div className="slider" id="sliderRegular"></div>
-              <br></br>
-              <div className="slider slider-primary" id="sliderDouble"></div>
-            </Col>
-          </Row>
-        </Container>
-      </div>
-    </>
-  );
-}
-
-export default BasicElements;
