--- bar-trender-front-end/src/views/index-sections/Images.js
@@ -1,41 +1,0 @@
-import React from "react";
-
-// reactstrap components
-import { Container, Row, Col } from "reactstrap";
-
-// core components
-
-function Images() {
-  return (
-    <>
-      <div className="section section-images">
-        <Container>
-          <Row>
-            <Col md="12">
-              <div className="hero-images-container">
-                <img
-                  alt="..."
-                  src={require("../../assets/img/hero-image-1.png")}
-                ></img>
-              </div>
-              <div className="hero-images-container-1">
-                <img
-                  alt="..."
-                  src={require("../../assets/img/hero-image-2.png")}
-                ></img>
-              </div>
-              <div className="hero-images-container-2">
-                <img
-                  alt="..."
-                  src={require("../../assets/img/hero-image-3.png")}
-                ></img>
-              </div>
-            </Col>
-          </Row>
-        </Container>
-      </div>
-    </>
-  );
-}
-
-export default Images;
